visibility: change default visibility to hidden for release build
